+++ ament_cmake_cppcheck/ament_cmake_cppcheck-extras.cmake
@@ -12,14 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# copied from ament_cppcheck/ament_cppcheck-extras.cmake
+# copied from ament_cmake_cppcheck/ament_cmake_cppcheck-extras.cmake
 
-find_package(ament_cmake_core QUIET REQUIRED)
 find_package(ament_cmake_test QUIET REQUIRED)
 
-set(ament_cppcheck_BIN "${ament_cppcheck_DIR}/../../../bin/ament_cppcheck")
+include("${ament_cmake_cppcheck_DIR}/ament_cppcheck.cmake")
 
-include("${ament_cppcheck_DIR}/ament_cppcheck.cmake")
-
-ament_register_extension("ament_lint_auto" "ament_cppcheck"
-  "ament_cppcheck_lint_hook.cmake")
+ament_register_extension("ament_lint_auto" "ament_cmake_cppcheck"
+  "ament_cmake_cppcheck_lint_hook.cmake")
