+++ ament_cppcheck/package.xml
@@ -2,19 +2,16 @@
 <package format="2">
   <name>ament_cppcheck</name>
   <version>0.0.0</version>
-  <description>The ability to perform static code analysis on C/C++ code using cppcheck in the ament buildsystem.</description>
+  <description>
+    The ability to perform static code analysis on C/C++ code using Cppcheck
+    and generate xUnit test result files.
+  </description>
   <maintainer email="dthomas@osrfoundation.org">Dirk Thomas</maintainer>
   <license>Apache License 2.0</license>
 
-  <buildtool_depend>ament_cmake_core</buildtool_depend>
-  <buildtool_depend>ament_cmake_python</buildtool_depend>
-  <buildtool_depend>ament_cmake_test</buildtool_depend>
+  <exec_depend>cppcheck</exec_depend>
 
-  <buildtool_export_depend>ament_cmake_core</buildtool_export_depend>
-  <buildtool_export_depend>ament_cmake_test</buildtool_export_depend>
-  <buildtool_export_depend>cppcheck</buildtool_export_depend>
-
-  <test_depend>ament_lint_cmake</test_depend>
-  <test_depend>ament_pep8</test_depend>
-  <test_depend>ament_pyflakes</test_depend>
+  <export>
+    <build_type>ament_python</build_type>
+  </export>
 </package>
